+++ src/Sphere.h
@@ -17,14 +17,14 @@
 
 class Sphere: public Primitive {
 	Vec3f position;
-	float radius;
+	double radius;
 	Vec3f calculateNormal(const Vec4f&) const;
 	void generateBoundingBox();
 	Mat4f inverseMatrixTranspose;
 public:
 	bool setTransformation(const Mat4f&);
-	Sphere(float, float, float, float, Mat4f&);
+	Sphere(double, double, double, double, Mat4f&);
 	virtual ~Sphere();
-	bool intersectiontest(Ray, float&, Primitive**) const;
+	bool intersectiontest(Ray, double&, Primitive**) const;
 };
 #endif /* SRC_SPHERE_H_ */
